\name{bSCDCsampler-package}
\alias{bSCDCsampler-package}
\alias{bSCDCsampler}
\docType{package}
\title{
\packageTitle{bSCDCsampler}
}
\description{
\packageDescription{bSCDCsampler}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bSCDCsampler}
\packageIndices{bSCDCsampler}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bSCDCsampler}

Maintainer: \packageMaintainer{bSCDCsampler}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
